+++ stories/assets/js/UNDP.js
@@ -12,3 +12,5 @@ UNDP.keyCode = {
 };
 
 window.UNDP = window.UNDP || UNDP;
+
+export const desktopView=window.innerWidth>1439;
