+++ stories/assets/js/navigation.js
@@ -1,112 +1,5 @@
+import {getData} from "./navigationData";
 var lang = 'english';
-
-const getData = () => {
-  $.ajax({
-    method: 'GET',
-    url: './js/navigationData.json',
-    dataType: 'json',
-    success(res) {
-      const resData = res.find((item) => item.language === lang);
-      console.log(resData);
-      // const menuData = resData.data.find((subItem) => subItem.id === id);
-      const megaWrapper = $('.mega-wrapper');
-      const mobileMegaWrapper = $('.mobile-mega-wrapper');
-      megaWrapper.empty();
-      mobileMegaWrapper.empty();
-      resData.data.forEach((item, index) => {
-        const submenus = item.submenus
-          .map((menu, index) => `<li id="${menu.link.id}-${item.id}" class="${
-            index === 0 ? 'active' : ''
-          }"><a href="#" class="sub-link">${menu.link.label}</a></li>`)
-          .join(' ');
-
-        const gridX = item.submenus
-          .map((menu, index) => `<div class="${
-            index === 0
-              ? 'grid-x sub-menu-content active-content'
-              : 'grid-x sub-menu-content'
-          }" data-id="${menu.link.id}-${item.id}">
-                      <div class="${
-  menu.image
-    ? 'cell mega-sub-description large-6'
-    : 'cell mega-sub-description large-10'
-}">
-                        <h3 class="mega-title">${menu.title}</h3>
-                        <p class="mega-description">${menu.description}</p>
-                        <div class="sub-sub-menu">
-                            <ul class="list-col-first">
-                            ${
-  menu.linksCol1.length > 0
-    ? menu.linksCol1
-      .map((link) => `<li><a class="cta__link cta--arrow" href="#">${link.label}</a></li>`)
-      .join(' ')
-    : ''
-}
-                            </ul>
-                            <ul class="list-col-second">
-                            ${
-  menu.linksCol2.length > 0
-    ? menu.linksCol2
-      .map((link) => `<li><a class="cta__link cta--arrow" href="#">${link.label}</a></li>`)
-      .join(' ')
-    : ''
-}
-                            </ul>
-                        </div>
-                      </div>
-                      ${
-  menu.image
-    ? `<div class="cell large-6 mega-image"><img src="${menu.image}" alt="Author Image" class="image"></div>`
-    : ''
-}
-                  </div>`)
-          .join(' ');
-
-        const renderHtml = `
-                    <div class="cell large-12 mega-nav-option" menu-item-id="${item.id}">
-                    <div class="grid-x">
-                        <div class="cell large-3">
-                          <ul class="submenu">
-                              ${submenus}
-                          </ul>
-                        </div>
-                        <div class="cell large-9">
-                          ${gridX}
-                        </div>
-                    </div>
-           </div>
-        `;
-        megaWrapper.append(renderHtml);
-
-        // render mobile mega navs
-        const renderMobileHtml = `
-        <div class="mobile-mega-content" data-mobile-id="${item.id}">
-          <h6 class="sub-heading">${item.title}</h6>
-          ${item.submenus
-    .map((submenu, index) => `<ul class="sub-sub-menus">
-                      <li>
-                        <span>${submenu.link.label}</span>
-                        <ul>
-                        ${[...submenu.linksCol1, ...submenu.linksCol2]
-    .map((link, index) => `<li>
-                            <a href="#">${link.label}</a>
-                          </li>`)
-    .join(' ')}
-                        </ul>
-                      </li>
-                </ul>`)
-    .join(' ')}
-        </div>
-        `;
-        mobileMegaWrapper.append(renderMobileHtml);
-      });
-    },
-    error(err) {
-      console.log(err);
-    },
-  });
-};
-
 
 export const init = (locale) => {
   const menuItem = $('.menu-item');
@@ -119,7 +12,6 @@ export const init = (locale) => {
   getData();
 
   menuItem.on('mouseenter', function () {
-    // megaNavOption.addClass('show-mega');
     const id = $(this).attr('id');
     $(document).find(`[menu-item-id='${id}']`).siblings().removeClass('show-mega');
     $(document).find(`[menu-item-id='${id}']`).addClass('show-mega');
